fix /r/translaki.md
It linked to the original trans act from 2002-06-28 instead of the most
recent version of the law, I noticed this when I started wondering
about the divorcing requirement still being there while equal marriage
law removed it. I think that this link is more useful for me as the
current version which came to force 2017-03-01.