r/flatpakassuperuser: flatpak is not root
Flatpak apps appear in MATE like they were running as root and this
issue says that it's incorrect information. I imagine that it can be
confusing to other people too and as it does confuse me, I imagine there
may be a reason for me to link it to other people too.