index: add sadly still necessary post from 2015
IIRC I was writing it in hope/fear of someone I know offline reading this
and that it might save me from pain such discussions tend to bring.

Then on 7ce2f89c5f4a65ffedc56eea3158142d0a9e459f I judged it as paranoia
and removed it (I think the blog has disclaimer about my opinions
possibly not being so static and that I may be a different person from
what I was at time of some typings).

Today then one channel at IRC / group at Telegram has been discussing
about me or my transness for hours thanks to this line not being common
sense, so I guess it's best to restore the message and maybe link people
to this commit message in the future if they are bothered by it.