r: add git.md to redirect to nice writeup on it
I originally saw this a few weeks ago when someone else linked me to it
and now I was asked for it by someone possibly thinking I had originally
linked into it and it seems so nice I want to store it somewhere.

And this somewhere is my most persistent bookmarks storage.