blog/[m]ppn: iOS pills, clicksaver for release day changes
and in the last section also mention emoji picker which isn't
particularly noteworthy, but it came up in the same discussion as lack
of pills and was unreported to upstream and as I was multitasking it was
two things in one.

I hope it not being that important in my opinion
opens to the reader for it being so last mention there.